Refactor @mergin/lib: - rename project access requests table to more sensible name - it is handling global access requests notonly projects - method getProjectAccessRequests is handling also global access requests (namespace) - rename to getAccessRequests
Cleanup @mergin/lib:
- remove unused BaseItem component
