--- web-app/packages/lib/src/common/components/base/BaseItem.vue
@@ -1,105 +1,0 @@
-<!--
-Based on template:
-Product Page: https://www.creative-tim.com/product/vuetify-material-dashboard
-Copyright 2019 Creative Tim (https://www.creative-tim.com)
-
-Modifications by:
-Copyright (C) Lutra Consulting Limited
-
-SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-MerginMaps-Commercial
--->
-
-<template>
-  <v-list-item
-    :href="href"
-    :rel="href && href !== '#' ? 'noopener' : undefined"
-    :target="href && href !== '#' ? '_blank' : undefined"
-    :to="item.to"
-    active-class="baseItemActiveColor"
-  >
-    <!--  TODO: VUE 3 - problematic upgrade, check this   -->
-    <template #prepend>
-      <v-list-item
-        v-if="text"
-        class="v-list-item__icon--text"
-        v-text="computedText"
-        :class="`${color}--text`"
-        icon
-      />
-
-      <v-list-item
-        v-else-if="item.icon || item.tablerIcon"
-        class="mr-2"
-        :class="`${color}--text`"
-        icon
-      >
-        <v-icon v-if="item.icon" :class="`${color}--text`" v-text="item.icon" />
-        <slot v-if="item.tablerIcon" name="icon"></slot>
-      </v-list-item>
-    </template>
-
-    <v-layout v-if="item.title || item.subtitle" :class="`${color}--text`">
-      <v-list-item-title v-text="item.title" />
-
-      <v-list-item-subtitle v-text="item.subtitle" />
-    </v-layout>
-  </v-list-item>
-</template>
-
-<script lang="ts">
-import { defineComponent, PropType } from 'vue'
-
-import { SideBarItemModel } from '@/modules/layout/types'
-// import Themeable from 'vuetify/lib/mixins/themeable'
-
-export default defineComponent({
-  name: 'BaseItem',
-
-  // mixins: [Themeable],
-
-  props: {
-    color: {
-      type: String,
-      default: 'white'
-    },
-    item: {
-      type: Object as PropType<SideBarItemModel>,
-      default: () => ({
-        href: undefined,
-        icon: undefined,
-        tablerIcon: false,
-        subtitle: undefined,
-        title: undefined,
-        to: undefined
-      })
-    },
-    text: {
-      type: Boolean,
-      default: false
-    }
-  },
-
-  computed: {
-    computedText() {
-      if (!this.item || !this.item.title) return ''
-
-      let text = ''
-
-      this.item.title.split(' ').forEach((val) => {
-        text += val.substring(0, 1)
-      })
-
-      return text
-    },
-    href() {
-      return this.item.href || (this.item.to ? undefined : '#')
-    }
-  }
-})
-</script>
-
-<style lang="scss" scoped>
-.baseItemActiveColor {
-  background-color: #f3f4f8;
-}
-</style>
