--- web-app/packages/lib/src/common/components/base/index.ts
@@ -1,5 +1,0 @@
-// Copyright (C) Lutra Consulting Limited
-//
-// SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-MerginMaps-Commercial
-
-export { default as BaseItem } from './BaseItem.vue'
