+++ web-app/packages/lib/src/modules/project/components/AccessRequestTable.vue
@@ -5,7 +5,7 @@ SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-MerginMaps-Commercial
 -->
 
 <template>
-  <project-access-request-table-template
+  <access-request-table-template
     :namespace="namespace"
     @accept-access-request-error="onAcceptAccessRequestError"
   />
@@ -16,14 +16,14 @@ import { AxiosError } from 'axios'
 import { mapActions } from 'pinia'
 import { defineComponent } from 'vue'
 
-import ProjectAccessRequestTableTemplate from './ProjectAccessRequestTableTemplate.vue'
+import AccessRequestTableTemplate from './AccessRequestTableTemplate.vue'
 
 import { getErrorMessage } from '@/common/error_utils'
 import { useNotificationStore } from '@/modules/notification/store'
 
 export default defineComponent({
-  name: 'ProjectAccessRequestTable',
-  components: { ProjectAccessRequestTableTemplate },
+  name: 'AccessRequestTable',
+  components: { AccessRequestTableTemplate },
   props: {
     namespace: {
       type: String,
