+++ web-app/packages/lib/src/modules/project/components/index.ts
@@ -11,7 +11,7 @@ export { default as FileIcon } from './FileIcon.vue'
 export { default as FolderDiff } from './FolderDiff.vue'
 export { default as PermissionInfo } from './PermissionInfo.vue'
 export { default as ProjectAccessRequests } from './ProjectAccessRequests.vue'
-export { default as ProjectAccessRequestTable } from './ProjectAccessRequestTable.vue'
+export { default as AccessRequestTable } from './AccessRequestTable.vue'
 export { default as ProjectForm } from './ProjectForm.vue'
 export { default as ProjectPermissionsTemplate } from './ProjectPermissionsTemplate.vue'
 export { default as ProjectShareDialog } from './ProjectShareDialog.vue'
@@ -20,6 +20,6 @@ export { default as ProjectShareButtonTemplate } from './ProjectShareButtonTempl
 export { default as ProjectsTable } from './ProjectsTable.vue'
 export { default as ProjectsTableDataLoader } from './ProjectsTableDataLoader.vue'
 export { default as UploadPanel } from './UploadPanel.vue'
-export { default as ProjectAccessRequestTableTemplate } from './ProjectAccessRequestTableTemplate.vue'
+export { default as AccessRequestTableTemplate } from './AccessRequestTableTemplate.vue'
 export { default as FileDetailSidebar } from './FileDetailSidebar.vue'
 export { default as VersionDetailSidebar } from './VersionDetailSidebar.vue'
